Move serialization files into dedicated module
