+++ evm/tests/empty_txn_list.rs
@@ -10,8 +10,7 @@ use keccak_hash::keccak;
 use log::info;
 use plonky2::field::goldilocks_field::GoldilocksField;
 use plonky2::plonk::config::PoseidonGoldilocksConfig;
-use plonky2::util::gate_serialization::default::DefaultGateSerializer;
-use plonky2::util::generator_serialization::default::DefaultGeneratorSerializer;
+use plonky2::util::serialization::{DefaultGateSerializer, DefaultGeneratorSerializer};
 use plonky2::util::timing::TimingTree;
 use plonky2_evm::all_stark::AllStark;
 use plonky2_evm::config::StarkConfig;
